Updated the Documentation Page
I added the gantt and requirements to the documentation via iframes as well as addded the figma via an image file. I also fixed the redirection to the home page from other pages because this used to lead to a blank page. I also removed several useless image files just to clear up some space.